Externalize exceptions
Primarily externalize exceptions. Began work on adding menu to Mainwindow.
