+++ emstencil/Exceptions.py
@@ -1,0 +1,43 @@
+#! /usr/bin/env python3
+"""
+ Program: Base custom exception classes.
+    Name: Andrew Dixon            File: exceptions.py
+    Date: 14 Sep 2024
+   Notes:
+
+    Copyright (C) 2024  Andrew Dixon
+
+    This program is free software: you can redistribute it and/or modify  it under the terms of the GNU
+    General Public License as published by the Free Software Foundation, either version 3 of the License,
+    or (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+    the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See theGNU General Public
+    License for more details.
+
+    You should have received a copy of the GNU General Public License along with this program.
+    If not, see <https://www.gnu.org/licenses/>.
+........1.........2.........3.........4.........5.........6.........7.........8.........9.........0.........1.........2.........3..
+"""
+
+class TemplateKeyValueMismatch(Exception):
+  """Exception for self.fields dictionary element mismatch"""
+  def __init__(self, source: dict, dest: dict) -> None:
+    self.missingKeys = list(set(dest).symmetric_difference(source))
+    self.message = f'{self.missingKeys} not in source and destination'
+    super().__init__(self.message)
+
+
+class TemplateKeyValueNull(Exception):
+  """Exception for when a dictionary value is None"""
+  def __init__(self, value: str) -> None:
+    self.key = value
+    self.message = f'Value for key: [{self.key}] is Null'
+    super().__init__(self.message)
+
+
+class AccessNullRowID(Exception):
+  """Exception for when rowID expected and it is null"""
+  def __init__(self, *args: object) -> None:
+    self.message = 'Attempted to use rowID without first setting it'
+    super().__init__(self.message)
